order-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
kafka-python
pydantic
pytest
httpx
pytest-asyncio

